HTML project made for a class
The website is about Circuit Dev and it's kind of like futureshop or
bestbuy but it only sells desktops or laptops.

Initially, I made the index in order to have a template that I can use
repeatedly for stable site repeatition. I used GIMP to create the design
for the navigation bar and also for the logo.

Then I put some more links to the navigation bar for further page
developments. I created a free shipping advertisment for the index.html
page so I could attract customers. The logo and main menu links to the
homepage index.html.

Then I created contact us, careers and about us pages. I used ID
selectors for About us so it can redirect to the part of the page where
the client is interested in.

Laptops and desktop sites are more interesting because they use php
scripts. If you go to laptops link, you didnt check a radio and
submitted it then, nothing will be processed. However, if you check on
of the products and checked it out, you will be asked for some
shipping/billing informations.